Let `Game` emit commands
* `Game` trait now let a game emit `Cmd`, which tells the game manager
  to refeed into the game a given message, with delay if necessary.

* This enables us to implement events that happens periodically on the
  server without being instigated by a client request.

* With that in mind, the server now can tell to the clients that a round
  is over; to synchronize their timers with the server or delay the
  start of a new round.

+++ server/api.rs
@@ -123,7 +123,7 @@ impl fmt::Display for RoomId {
 }
 
 pub type Size = f32;
-pub type Score = i32;
+pub type Score = u16;
 pub type Color = String;
 pub type Point = (i32, i32);
